I have refactored my sign in code to conform to MVVM
