+++ Paws4Me/Paws4Me/Show and Validate Login/View Models/SignInViewModel.swift
@@ -1,0 +1,27 @@
+//
+//  ViewModels.swift
+//  Paws4Me
+//
+//  Created by Kirsty-Lee Walker on 2022/03/24.
+//
+
+import Foundation
+
+protocol Validation {
+    func isValidCredentials(username: String, password: String) -> Bool
+}
+
+class SignInViewModel: Validation {
+
+    // MARK: - Vars/Lets
+    private var isLoggedIn = false
+
+    // MARK: - Functions
+    func isValidCredentials(username: String, password: String) -> Bool {
+        if username == "Admin" && password == "TestPass123" {
+            return true
+        } else {
+            return false
+        }
+    }
+}
